Module_25 added fragments, refactoring
