+++ app/src/main/java/com/dascamp/myapp/DetailsFragment.kt
@@ -1,24 +1,33 @@
 package com.dascamp.myapp
 
 import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.dascamp.myapp.databinding.ActivityDetailsBinding
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.dascamp.myapp.databinding.FragmentDetailsBinding
 
-class DetailsActivity : AppCompatActivity() {
+class DetailsFragment : Fragment() {
 
-    private var detailBinding: ActivityDetailsBinding? = null
+    private var detailBinding: FragmentDetailsBinding? = null
+    private val binding get() = detailBinding
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        detailBinding = FragmentDetailsBinding.inflate(inflater, container, false)
+        return binding?.root
+    }
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        detailBinding = ActivityDetailsBinding.inflate(layoutInflater)
-        setContentView(detailBinding!!.root)
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
 
         setFilmsDetails()
     }
 
     private fun setFilmsDetails() {
 
-        val film = intent.extras?.get("film") as Film
+        val film = arguments?.get("film") as Film
 
         detailBinding?.detailsToolbar?.title = film.title
         detailBinding?.detailsPoster?.setImageResource(film.poster)
