+++ app/src/main/res/layout/activity_main.xml
@@ -26,79 +26,12 @@
 
     </com.google.android.material.appbar.AppBarLayout>
 
-    <androidx.cardview.widget.CardView
-        android:id="@+id/top_bar"
+    <FrameLayout
+        android:id="@+id/fragment_placeholder"
         android:layout_width="match_parent"
-        android:layout_height="@dimen/cv2_height"
-        android:layout_margin="@dimen/cv2_margin"
-        android:background="@color/dark_violet"
-        android:elevation="@dimen/elevation"
-        app:cardCornerRadius="@dimen/card_corner_radius"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/app_bar_layout">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="horizontal">
-
-            <ImageView
-                android:id="@+id/ivPoster1"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_weight="@dimen/iv_weight"
-                android:contentDescription="@string/description"
-                android:scaleType="centerCrop"
-                android:src="@drawable/mermaid" />
-
-            <ImageView
-                android:id="@+id/ivPoster2"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_weight="@dimen/iv_weight"
-                android:contentDescription="@string/description"
-                android:scaleType="centerCrop"
-                android:src="@drawable/puss" />
-
-            <ImageView
-                android:id="@+id/ivPoster3"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_weight="@dimen/iv_weight"
-                android:contentDescription="@string/description"
-                android:scaleType="centerCrop"
-                android:src="@drawable/trolls3" />
-
-        </LinearLayout>
-    </androidx.cardview.widget.CardView>
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/main_recycler"
-        android:layout_width="0dp"
         android:layout_height="0dp"
-        android:layout_margin="16dp"
-        android:padding="8dp"
         app:layout_constraintBottom_toTopOf="@+id/bottom_navigation"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/disclaimer"
-        tools:listitem="@layout/film_item" />
-
-    <TextView
-        android:id="@+id/disclaimer"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/tv_marginTop"
-        android:layout_marginEnd="@dimen/tv_marginEnd"
-        android:autoLink="web"
-        android:linksClickable="true"
-        android:minHeight="@dimen/tv_minHeight"
-        android:text="@string/disclaimer"
-        android:textColor="@color/white"
-        android:textStyle="italic"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/top_bar" />
+        app:layout_constraintTop_toBottomOf="@+id/app_bar_layout" />
 
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:id="@+id/bottom_navigation"
