+++ app/src/main/res/layout/fragment_details.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".DetailsActivity">
+    tools:context=".DetailsFragment">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/app_bar"
@@ -27,9 +27,9 @@
                 android:id="@+id/details_poster"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
+                android:background="@color/design_default_color_primary"
                 android:fitsSystemWindows="true"
                 android:scaleType="centerCrop"
-                android:background="@color/design_default_color_primary"
                 android:src="@drawable/ic_launcher_foreground"
                 app:layout_collapseMode="parallax" />
 
@@ -50,11 +50,11 @@
         android:layout_height="wrap_content"
         android:layout_margin="16dp"
         android:backgroundTint="@color/teal_700"
+        android:contentDescription="@string/floating_action_button"
+        android:tint="@android:color/white"
         app:layout_anchor="@id/app_bar"
         app:layout_anchorGravity="bottom|right"
-        app:srcCompat="@drawable/ic_baseline_share_24"
-        android:tint="@android:color/white"
-        android:contentDescription="@string/floating_action_button" />
+        app:srcCompat="@drawable/ic_baseline_share_24" />
 
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
